move the upgrade-body-FIN test to its own file
node-http.test.ts has an unrelated pre-existing failure
('request via http proxy, issue#4295' ECONNREFUSED) in the
gate environment.
